+++ tilematrix/__init__.py
@@ -2,4 +2,4 @@
 
 from ._tilepyramid import TilePyramid
 from ._tile import Tile
-from ._funcs import clip_geometry_to_srs_bounds
+from ._funcs import clip_geometry_to_srs_bounds, Shape, Bounds, TileIndex