+++ tilematrix/_tilepyramid.py
@@ -2,10 +2,9 @@
 
 from shapely.prepared import prep
 import math
-from rasterio.crs import CRS
 
 from . import _funcs
-from ._conf import PYRAMID_PARAMS, ROUND, DELTA
+from ._conf import ROUND
 from ._tile import Tile
 
 
@@ -26,43 +25,15 @@ class TilePyramid(object):
 
     def __init__(self, grid_definition, tile_size=256, metatiling=1):
         """Initialize TilePyramid."""
-        if isinstance(grid_definition, dict):
-            self.type = "custom"
-            if "shape" not in grid_definition:
-                raise AttributeError
-            self._shape = _funcs.Shape(*grid_definition["shape"])
-            self.bounds = _funcs.Bounds(*grid_definition["bounds"])
-            # verify that shape aspect ratio fits bounds apsect ratio
-            _verify_shape_bounds(shape=self._shape, bounds=self.bounds)
-            self.left, self.bottom, self.right, self.top = self.bounds
-            self.is_global = grid_definition.get("is_global", False)
-            if all([i in grid_definition for i in ["proj", "epsg"]]):
-                raise ValueError("either 'epsg' or 'proj' are allowed.")
-            if "epsg" in grid_definition:
-                self.crs = CRS().from_epsg(grid_definition["epsg"])
-                self.srid = grid_definition["epsg"]
-            elif "proj" in grid_definition:
-                self.crs = CRS().from_string(grid_definition["proj"])
-                self.srid = None
-            else:
-                raise AttributeError("either 'epsg' or 'proj' is required")
-        else:
-            if grid_definition not in PYRAMID_PARAMS:
-                raise ValueError(
-                    "WMTS tileset '%s' not found. Use one of %s" % (
-                        grid_definition, PYRAMID_PARAMS.keys()))
-            if metatiling not in (1, 2, 4, 8, 16):
-                raise ValueError("metatling must be one of 1, 2, 4, 8, 16")
-            self.type = grid_definition
-            self._shape = _funcs.Shape(
-                *PYRAMID_PARAMS[grid_definition]["shape"])
-            self.bounds = _funcs.Bounds(
-                *PYRAMID_PARAMS[grid_definition]["bounds"])
-            self.left, self.bottom, self.right, self.top = self.bounds
-            self.is_global = PYRAMID_PARAMS[grid_definition]["is_global"]
-            self.srid = PYRAMID_PARAMS[grid_definition]["epsg"]
-            self.crs = CRS().from_epsg(self.srid)
-
+        # get source grid parameters
+        self.grid = _funcs.GridDefinition(
+            grid_definition, tile_size=tile_size, metatiling=metatiling)
+        self.type = self.grid.type
+        self.bounds = self.grid.bounds
+        self.left, self.bottom, self.right, self.top = self.bounds
+        self.crs = self.grid.crs
+        self.srid = self.grid.srid
+        self.is_global = self.grid.is_global
         self.metatiling = metatiling
         # size in pixels
         self.tile_size = tile_size
@@ -87,7 +58,8 @@ class TilePyramid(object):
 
         - zoom: zoom level
         """
-        width = int(math.ceil(self._shape.width * 2**(zoom) / self.metatiling))
+        width = int(math.ceil(
+            self.grid.shape.width * 2**(zoom) / self.metatiling))
         return 1 if width < 1 else width
 
     def matrix_height(self, zoom):
@@ -97,7 +69,7 @@ class TilePyramid(object):
         - zoom: zoom level
         """
         height = int(
-            math.ceil(self._shape.height * 2**(zoom) / self.metatiling))
+            math.ceil(self.grid.shape.height * 2**(zoom) / self.metatiling))
         return 1 if height < 1 else height
 
     def tile_x_size(self, zoom):
@@ -122,7 +94,7 @@ class TilePyramid(object):
 
         - zoom: zoom level
         """
-        matrix_pixel = 2**(zoom) * self.tile_size * self._shape.width
+        matrix_pixel = 2**(zoom) * self.tile_size * self.grid.shape.width
         tile_pixel = self.tile_size * self.metatiling
         return matrix_pixel if tile_pixel > matrix_pixel else tile_pixel
 
@@ -132,7 +104,7 @@ class TilePyramid(object):
 
         - zoom: zoom level
         """
-        matrix_pixel = 2**(zoom) * self.tile_size * self._shape.height
+        matrix_pixel = 2**(zoom) * self.tile_size * self.grid.shape.height
         tile_pixel = self.tile_size * self.metatiling
         return matrix_pixel if tile_pixel > matrix_pixel else tile_pixel
 
@@ -221,31 +193,25 @@ class TilePyramid(object):
             "MultiPoint", "GeometryCollection"
         ):
             prepared_geometry = prep(
-                _funcs.clip_geometry_to_srs_bounds(geometry, self)
-            )
+                _funcs.clip_geometry_to_srs_bounds(geometry, self))
             for tile in self.tiles_from_bbox(geometry, zoom):
                 if prepared_geometry.intersects(tile.bbox()):
                     yield tile
         else:
             raise ValueError("no valid geometry: %s" % geometry.type)
 
+    def __eq__(self, other):
+        if isinstance(other, self.__class__):
+            return all([
+                self.grid == other.grid,
+                self.tile_size == other.tile_size,
+                self.metatiling == other.metatiling])
+        else:
+            return False
 
-def _verify_shape_bounds(shape, bounds):
-    """Verify that shape corresponds to bounds apect ratio."""
-    shape = _funcs.Shape(*map(float, shape))
-    bounds = _funcs.Bounds(*map(float, bounds))
-    shape_ratio = shape.width / shape.height
-    bounds_ratio = (bounds.right - bounds.left) / (bounds.top - bounds.bottom)
-    if abs(shape_ratio - bounds_ratio) > DELTA:
-        min_length = min([
-            (bounds.right - bounds.left) / shape.width,
-            (bounds.top - bounds.bottom) / shape.height
-        ])
-        proposed_bounds = _funcs.Bounds(
-            bounds.left,
-            bounds.bottom,
-            bounds.left + shape.width * min_length,
-            bounds.bottom + shape.height * min_length)
-        raise ValueError(
-            "shape ratio (%s) must equal bounds ratio (%s); try %s" % (
-                shape_ratio, bounds_ratio, proposed_bounds))
+    def __ne__(self, other):
+        return not self.__eq__(other)
+
+    def __repr__(self):
+        return 'TilePyramid(%s, tile_size=%s, metatiling=%s)' % (
+            self.type, self.tile_size, self.metatiling)